rename route /app to /connect-third-party-apps and handle Strava connection
